+++ inference_experimental/inference_exp/models/grounding_dino/grounding_dino_torch.py
@@ -19,14 +19,14 @@ from torchvision.ops import box_convert
 
 try:
     from groundingdino.util.inference import load_model, predict
-except ImportError:
+except ImportError as import_error:
     raise MissingDependencyError(
         f"Could not import GroundingDino model - this error means that some additional dependencies "
         f"are not installed in the environment. If you run the `inference` library directly in your Python "
         f"program, make sure the following extras of the package are installed: `grounding-dino`."
         f"If you see this error using Roboflow infrastructure, make sure the service you use does support the model. "
         f"You can also contact Roboflow to get support."
-    )
+    ) from import_error
 
 
 DEFAULT_CONFIG_URL = "https://raw.githubusercontent.com/roboflow/GroundingDINO/main/groundingdino/config/GroundingDINO_SwinT_OGC.py"
