+++ inference_experimental/inference_exp/models/l2cs/l2cs_onnx.py
@@ -17,7 +17,7 @@ from torchvision import transforms
 
 try:
     import onnxruntime
-except ImportError:
+except ImportError as import_error:
     raise MissingDependencyError(
         f"Could not import L2CS model with ONNX backend - this error means that some additional dependencies "
         f"are not installed in the environment. If you run the `inference` library directly in your Python "
@@ -28,7 +28,7 @@ except ImportError:
         f"\t* `onnx-jp6-cu126` - for running on Jetson with Jetpack 6\n"
         f"If you see this error using Roboflow infrastructure, make sure the service you use does support the model. "
         f"You can also contact Roboflow to get support."
-    )
+    ) from import_error
 
 
 DEFAULT_GAZE_MAX_BATCH_SIZE = 8
