+++ inference_experimental/inference_exp/models/mediapipe_face_detection/face_detection.py
@@ -12,14 +12,14 @@ from inference_exp.models.common.model_packages import get_model_package_content
 try:
     import mediapipe as mp
     from mediapipe.tasks.python.components.containers import Detection
-except ImportError:
+except ImportError as import_error:
     raise MissingDependencyError(
         f"Could not import face detection model from MediaPipe - this error means that some additional dependencies "
         f"are not installed in the environment. If you run the `inference` library directly in your Python "
         f"program, make sure the following extras of the package are installed: `mediapipe`."
         f"If you see this error using Roboflow infrastructure, make sure the service you use does support the model. "
         f"You can also contact Roboflow to get support."
-    )
+    ) from import_error
 
 
 class MediaPipeFaceDetector(
