+++ inference_experimental/inference_exp/models/yolov5/yolov5_object_detection_onnx.py
@@ -27,7 +27,7 @@ from inference_exp.models.yolov5.nms import run_nms_yolov5
 
 try:
     import onnxruntime
-except ImportError:
+except ImportError as import_error:
     raise MissingDependencyError(
         f"Could not import YOLOv5 model with ONNX backend - this error means that some additional dependencies "
         f"are not installed in the environment. If you run the `inference` library directly in your Python "
@@ -38,7 +38,7 @@ except ImportError:
         f"\t* `onnx-jp6-cu126` - for running on Jetson with Jetpack 6\n"
         f"If you see this error using Roboflow infrastructure, make sure the service you use does support the model. "
         f"You can also contact Roboflow to get support."
-    )
+    ) from import_error
 
 
 class YOLOv5ForObjectDetectionOnnx(
